LPS-36216 Replace getXYByUuidAndGroupId methods to fetchXYByUuidAndGroupId now that we have them for all models

+++ portal-impl/src/com/liferay/portlet/journal/lar/JournalFolderStagedModelDataHandler.java
@@ -44,10 +44,12 @@ public class JournalFolderStagedModelDataHandler
 		throws PortalException, SystemException {
 
 		JournalFolder folder =
-			JournalFolderLocalServiceUtil.getJournalFolderByUuidAndGroupId(
+			JournalFolderLocalServiceUtil.fetchJournalFolderByUuidAndGroupId(
 				uuid, groupId);
 
-		JournalFolderLocalServiceUtil.deleteFolder(folder);
+		if (folder != null) {
+			JournalFolderLocalServiceUtil.deleteFolder(folder);
+		}
 	}
 
 	@Override
